pyproject.toml
[tool.poetry]
name = "wozniak-coffee-test"
version = "0.1.0"
description = "Repository developed to complete Wozniak's coffe test"
authors = ["Breno U. de Angelo <brenodeangelo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.33.0"
termcolor = "^2.4.0"
tenacity = "^8.3.0"
gradio = "^4.36.1"
openai-whisper = {git = "https://github.com/openai/whisper.git", rev = "ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
